+++ tests/basicboot.rs
@@ -1,7 +1,7 @@
 #![no_std]
 #![no_main]
 #![feature(custom_test_frameworks)]
-#![test_runner(crate::testexec)]
+#![test_runner(libertyos_kernel::testexec)]
 #![reexport_test_harness_main = "testmain"]
 
 use core::panic::PanicInfo;
@@ -14,10 +14,10 @@ pub extern "C" fn _start() -> !
 }
 
 
-fn testexec(tests: &[&dyn Fn()])
-{
-	unimplemented!();
-}
+//fn testexec(tests: &[&dyn Fn()])
+//{
+//	unimplemented!();
+//}
 
 #[panic_handler]
 fn panic(info: &PanicInfo) -> !